+++ backend/wisdom_edu_api/asgi.py
@@ -2,15 +2,16 @@
 
 import os
 
-from channels.routing import ProtocolTypeRouter, URLRouter
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "wisdom_edu_api.settings")
+
 from django.core.asgi import get_asgi_application
+
+django_asgi_app = get_asgi_application()
+
+from channels.routing import ProtocolTypeRouter, URLRouter
 
 from ai_services.auth import QueryStringJWTAuthMiddlewareStack
 from ai_services.routing import websocket_urlpatterns
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "wisdom_edu_api.settings")
-
-django_asgi_app = get_asgi_application()
 
 application = ProtocolTypeRouter(
     {